pyproject.toml
[project]
name = "youtube-fastapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.20.0",
    "alembic>=1.14.0",
    "async-sqlmodel>=0.1.3",
    "fastapi[standard]>=0.115.5",
    "passlib>=1.7.4",
    "pyjwt>=2.10.1",
]
